:x: Add failing test describing game update function.
I've also moved the game instance creation to a session wide fixture.
I did that because the current game initialization create a showbase instance of which there can be only one.
Probably this reveals that the Game object should not be tested directly and that all its functions should be
refactored into separate modules.